requirements.txt
Django
psycopg2
djangorestframework
drf-yasg
django-cors-headers